Full Library Conversion and Documentation
