+++ Cargo.toml
@@ -1,8 +1,6 @@
 [package]
 name = "giftbox"
-version = "0.1.0"
+version = "0.1.1"
 edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
